move board to separate file
